\name{ICMg.combined.iteration}
\alias{ICMg.combined.iteration}
\title{ ICMG.combined.iteration }
\description{
  Internal iteration function for ICMg tools.
}
\usage{
ICMg.combined.iteration(L, X, Niter, N, M, Lindices, Nindices, D, C, z, w, n, m, q, E, alpha, beta, pm0, V0, V, convl, convn)
}
\author{ Juuso Parkkinen }
\note{ Called automatically by ICMg.combined.wrapper. See argument details from ICMg.combined.sampler!
}
\seealso{ \code{\link{ICMg.combined.sampler}},\code{\link{ICMg.combined.wrapper}}}

\keyword{ internal }
\keyword{ iteration }
