\name{ICMg.combined.wrapper}
\alias{ICMg.combined.wrapper}
\title{ ICMg.combined.wrapper }
\description{ Internal function for ICMg tools.
}
\usage{ ICMg.combined.wrapper(L, X, Niter, N, M, Lindices, Nindices,D, C, z, w, n, m, q, E,alpha, beta, pm0, V0, V,convl, convn, C.boost) }
\author{ Juuso Parkkinen }
\note{ Called automatically by ICMg.combined.sampler. See argument details there.
}
\seealso{ \code{\link{ICMg.combined.sampler}}}
\keyword{ internal }
