\name{ICMg.links.iteration}
\alias{ICMg.links.iteration}
\title{ ICMg.links.iteration }
\description{
  Internal iteration function for ICMg tools.
}
\usage{
ICMg.links.iteration(L, Niter, N, M, Lindices, C, z, q, n, alpha, beta, conv)
}
\author{ Juuso Parkkinen }
\note{ Called automatically by ICMg.links.wrapper. See argument details from ICMg.links.sampler!
}
\seealso{ \code{\link{ICMg.links.sampler}},\code{\link{ICMg.links.wrapper}}}

\keyword{ internal }
\keyword{ iteration }
