\name{ICMg.links.iteration.Weights}
\alias{ICMg.links.iteration.Weights}
\title{ ICMg.links.iteration.Weights }
\description{
  Internal iteration function for wMICA tools.
}
\usage{
ICMg.links.iteration.Weights(L, Niter, N, M, Lindices, C, z, q, n, alpha, beta, conv)
}
\author{ Christoph }
\note{ Called automatically by ICMg.links.wrapper.Weights. See argument details from ICMg.links.sampler!
}
\seealso{ \code{\link{ICMg.links.sampler.SeedsandWeights}},\code{\link{ICMg.links.wrapper.Weights}}}

\keyword{ internal }
\keyword{ iteration }
