\name{ICMg.links.wrapper}
\alias{ICMg.links.wrapper}
\title{ ICMg.links.wrapper }
\description{ Internal function for wMICA tools.
}
\usage{ ICMg.links.wrapper(L, Niter, N, M, Lindices,C, z, q, n,alpha, beta, conv, C.boost) }
\author{ Juuso Parkkinen }
\note{ Called automatically by ICMg.links.sampler. See argument details there.
}
\seealso{ \code{\link{ICMg.links.sampler}}}
\keyword{ internal }
