\name{ICMg.links.wrapper.Weights}
\alias{ICMg.links.wrapper.Weights}
\title{ ICMg.links.wrapper.Weights }
\description{ Internal function for wMICA tools.
}
\usage{ ICMg.links.wrapper(L, Niter, N, M, Lindices,C, z, q, n,alpha, beta, conv, C.boost) }
\author{ Christoph }
\note{ Called automatically by ICMg.links.sampler.SeedsAndWeights. See argument details there.
}
\seealso{ \code{\link{ICMg.links.sampler.SeedsAndWeights}}}
\keyword{ internal }
