\name{MICASample}
\alias{MICASample}
\docType{data}
\title{Sample from Transcriptome Study}
\description{
  Sample Data from the Hybrid Mouse Diversity Panel
}
\usage{data(MICASample)}
\format{ Singe Matrix with 200 genes and 25 strains of mice
}
\details{
  Number of genes: 200
}
\source{
  WIP
}
\references{
	
}
\examples{
data(MICASample)
}
\keyword{datasets}
